wir_build: rename translate_* helpers to concern-based names
The `translate_` prefix was only a historical artifact from splitting
translate.rs. Each helper module's name now directly reflects its concern,
matching the existing convention in wir_build/ (context.rs, functions.rs,
types.rs, component_plan.rs).

Renames:
- translate_value_copy.rs → value_copy.rs
- translate_ops.rs        → primitive_ops.rs
- translate_calls.rs      → calls.rs
- translate_canonical.rs  → canonical_abi.rs
- translate_match.rs      → pattern_match.rs

translate.rs itself is kept as the driver module, since it holds the
FunctionTranslator struct, LabelEntry, the pass entry points
(register_closure_wrappers, translate_function_bodies), and the core
translate_expr/stmt/block dispatch.
